requirements.txt
asyncio==3.4.3
discord.py==1.6.0
nekos.py==1.0.3
nest-asyncio==1.4.0
wikipedia==1.4.0
mcstatus==4.0.0
beautifulsoup4==4.9.1
youtube-dl==2021.3.25
google-api-python-client==1.12.3
PyNaCl==1.4.0
pymongo==3.11.0
dnspython==1.16
requests==2.25.1
ffmpeg==1.4
attrs==20.3.0
psutil==5.8.0
